Core no longer dependent on SDH
