+++ src/Session/DTMF.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from "events";
 
 import { C } from "../Constants";
-import { fromBodyObj, IncomingRequest } from "../Core/messages";
+import { IncomingRequest } from "../Core/messages";
 import { SessionStatus, TypeStrings } from "../Enums";
 import { Exceptions } from "../Exceptions";
 import { Logger } from "../LoggerFactory";
@@ -108,7 +108,7 @@ export class DTMF extends EventEmitter {
     if (this.owner.session) {
       const request = this.owner.session.info(undefined, {
         extraHeaders,
-        body: fromBodyObj(body)
+        body: Utils.fromBodyObj(body)
       });
       this.owner.emit("dtmf", request.message, this);
       return;
